+++ components/portfolio/portfolioSections/SimpleDividentScheduleSection/index.tsx
@@ -1,4 +1,4 @@
-import CommonFont from '@/components/Font';
+import CommonFont from '@/components/common/Font';
 import { ScheduleList } from '@/components/List/ScheduleList';
 import Section from '@/components/Section';
 import CommonBar from '@/components/common/Bar';
